requirements.txt
google-generativeai==0.3.1
gradio==2.0.7
python-dotenv==0.19.1